fix: handling of time
If Kibana is set to use absolute time, the code would fail because
the Kibana isoformat includes a 'Z' which the Python code does not
parse.  The fix also adjusts the manner that time format errors
are handled; instead of returning an ambiguous None...an exception
is thrown that then causes an explicit 500 error.