feat: change method for simplifying categorical _colorize
Instead of coloring only the first 1000 items (and leaving the others as
Other) take advantage of the fact that often we are assigning the same
color to multiple categories.  There are only 256 colors so when the
number of unique categories is large rather than given datashader tens
of thousands of categories, we can just replace categories with the
colors.  Thus the memory used by Datashader will never exceed
h*w*n_unique_colors.